removed converter code from stnode
