+++ src/roman_datamodels/stnode.py
@@ -81,34 +81,6 @@ def _get_schema_for_property(schema, attr):
                 return subsubschema
     return {}
 
-# class NodeIterator:
-#     """
-#     An iterator for a node which flattens the hierachical structure
-#     """
-#     def __init__(self, node):
-#         self.key_stack = []
-#         self.iter_stack = [iter(node._data.items())]
-
-#     def __iter__(self):
-#         return self
-
-#     def __next__(self):
-#         while self.iter_stack:
-#             try:
-#                 key, val = next(self.iter_stack[-1])
-#             except StopIteration:
-#                 self.iter_stack.pop()
-#                 if self.iter_stack:
-#                     self.key_stack.pop()
-#                 continue
-
-#             if isinstance(val, dict):
-#                 self.key_stack.append(key)
-#                 self.iter_stack.append(iter(val.items()))
-#             else:
-#                 return '.'.join(self.key_stack + [key])
-
-#         raise StopIteration
 
 class DNode(UserDict):
 
@@ -306,308 +278,4 @@ class TaggedObjectNodeConverter(Converter):
     def from_yaml_tree(self, node, tag, ctx):
         return (node)
 
-# class TopTaggedObjectNodeConverter(TaggedObjectNodeConverter):
-#     """
-#     This class is inteded to be used for all top level tags,
-#     e.g., those associated with data products handled by 
-#     pipeline steps.
 
-#     It is implicitly assumed that all instances will have a meta.filename
-#     attribute
-#     """
-
-#     def crds_observatory(self):
-#         """
-#         Get CRDS observatory code for this model.
-
-#         Returns
-#         -------
-#         str
-#         """
-#         return "roman"
-
-#     def get_crds_parameters(self):
-#         """
-#         Get parameters used by CRDS to select references for this model.
-
-#         Returns
-#         -------
-#         dict
-#         """
-#         return {
-#             key: val for key, val in self.to_flat_dict(include_arrays=False).items()
-#             if isinstance(val, (str, int, float, complex, bool))
-#         }
-
-#     def save(self, path, dir_path=None, *args, **kwargs):
-#         pass
-
-###################################
-#
-# Roman section
-#
-###################################
-
-class WfiScienceRaw(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/wfi_science_raw-1.0.0"
-
-class WfiScienceRawConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/wfi_science_raw-*"]
-    types = ["roman_datamodels.stnode.WfiScienceRaw"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return WfiScienceRaw(node)
-
-
-class WfiImage(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/wfi_image-1.0.0"
-
-class WfiImageConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/wfi_image-*"]
-    types = ["roman_datamodels.stnode.WfiImage"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        del obj._data['meta']['cal_step']
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        objnode = WfiImage(node)
-        objnode.meta['cal_step'] = objnode.meta.calstatus
-        return objnode
-
-class WfiMode(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/wfi_mode-1.0.0"
-
-class WfiModeConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/wfi_mode-*"]
-    types = ["roman_datamodels.stnode.WfiMode"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return WfiMode(node)
-
-class Exposure(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/exposure-1.0.0"
-
-class ExposureConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/exposure-*"]
-    types = ["roman_datamodels.stnode.Exposure"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Exposure(node)
-
-class Wfi(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/wfi-1.0.0"
-
-class WfiConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/wfi-*"]
-    types = ["roman_datamodels.stnode.Wfi"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Wfi(node)
-
-class Program(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/program-1.0.0"
-
-class ProgramConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/program-*"]
-    types = ["roman_datamodels.stnode.Program"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Program(node)
-
-class Observation(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/observation-1.0.0"
-
-class ObservationConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/observation-*"]
-    types = ["roman_datamodels.stnode.Observation"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Observation(node)
-
-class Ephemeris(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/ephemeris-1.0.0"
-
-class EphemerisConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/ephemeris-*"]
-    types = ["roman_datamodels.stnode.Ephemeris"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Ephemeris(node)
-
-class Visit(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/visit-1.0.0"
-
-class VisitConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/visit-*"]
-    types = ["roman_datamodels.stnode.Visit"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Visit(node)
-
-class Photometry(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/photometry-1.0.0"
-
-class PhotometryConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/photometry-*"]
-    types = ["roman_datamodels.stnode.Photometry"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Photometry(node)
-
-class Coordinates(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/coordinates-1.0.0"
-
-class CoordinatesConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/coordinates-*"]
-    types = ["roman_datamodels.stnode.Coordinates"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Coordinates(node)
-
-class Aperture(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/aperture-1.0.0"
-
-class ApertureConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/aperture-*"]
-    types = ["roman_datamodels.stnode.Aperture"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Aperture(node)
-
-class Pointing(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/pointing-1.0.0"
-
-class PointingConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/pointing-*"]
-    types = ["roman_datamodels.stnode.Pointing"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Pointing(node)
-
-class Target(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/target-1.0.0"
-
-class TargetConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/target-*"]
-    types = ["roman_datamodels.stnode.Target"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Target(node)
-
-class VelocityAberration(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/velocity_aberration-1.0.0"
-
-class VelocityAberrationConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/velocity_aberration-*"]
-    types = ["roman_datamodels.stnode.VelocityAberration"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return VelocityAberration(node)
-
-class Wcsinfo(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/wcsinfo-1.0.0"
-
-class WcsinfoConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/wcsinfo-*"]
-    types = ["roman_datamodels.stnode.Wcsinfo"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Wcsinfo(node)
-
-class Guidestar(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/guidestar-1.0.0"
-
-class GuidestarConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/guidestar-*"]
-    types = ["roman_datamodels.stnode.Guidestar"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Guidestar(node)
-
-class Program(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/program-1.0.0"
-
-class ProgramConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/program-*"]
-    types = ["roman_datamodels.stnode.Program"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Program(node)
-
-class Calstatus(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/calstatus-1.0.0"
-
-class CalstatusConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/calstatus-*"]
-    types = ["roman_datamodels.stnode.Calstatus"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return Calstatus(node)
-
-class FlatRef(TaggedObjectNode):
-    _tag = "tag:stsci.edu:datamodels/roman/reference_files/flat-1.0.0"
-
-class FlatRefConverter(TaggedObjectNodeConverter):
-    tags = ["tag:stsci.edu:datamodels/roman/reference_files/flat-*"]
-    types = ["roman_datamodels.stnode.FlatRef"]
-
-    def to_yaml_tree(self, obj, tags, ctx):
-        return obj._data
-
-    def from_yaml_tree(self, node, tag, ctx):
-        return FlatRef(node)
